claim_verification/requirements.txt
# Core dependencies
langchain>=0.3.0
langchain-anthropic>=0.2.0
langgraph>=0.2.0
langchain-community>=0.3.0
duckduckgo-search>=6.0.0  # Required for DuckDuckGo search tool
crawl4ai>=0.4.0
pydantic>=2.0
python-dotenv>=1.0.0

# Async support
asyncio
aiohttp

# Utils
typing-extensions>=4.0.0
